Implemented Object types for simpleobjects and basic sticky platforms
I have added the objecttype enum into simplegameobject. This involves several getters and setters that can determine the state of this enum. Along with this is a detector for collision with objects that have the lethal enum, killing the player upon contact.

I have also implementied rudimentary sticky platforms by detecting collisions between the player and weighted platform, then making the player update their position based on a set displacement. However, there are many bugs and the system is quite flawed, and will likely be changed alot.